rename migrations base test file
